Update Controller Salle de sport

+++ src/main/java/com/example/sportzone/Repository/ServiceRepository.java
@@ -3,5 +3,6 @@ package com.example.sportzone.Repository;
 import com.example.sportzone.entity.ServiceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<ServiceEntity,Integer> {
+public interface ServiceRepository extends JpaRepository<ServiceEntity, Long> {
+    // Custom queries can be added here if needed
 }
